Rename staking tokens to `*ZEAL`. Rationale:
If all tokenizations of ZEC maintain the invariant that <something>ZEC<something> represents a token intended to have a value pegged to mainnet ZEC, it will improve usability and comprehension for users across all ecosystems around ZEC. Any token with any freedom in exchange rates should have a different name. I picked "ZEAL" whimsically, but not particularly attached to it. ;-)